Move cmv and test_cmv to appropriate folders
